+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ## v0.1.3 - 2024-07-29
 
-- Fix leaking error on serializing `lat_max`, `lon_min` and `lon_max`
+- Fix serializing `lat_max`, `lon_min` and `lon_max`
 - Fix to relax `dump` and `dumps` args type
 
 ## v0.1.2 - 2024-07-27